Scan block throttle (#7)
* More compiler warnings activated

* Fixed gcc warnings. Added gcc as mac compiler.

* Added block creation throttling

* Set gcc warning level to pedantic

* Added more compiler warning flags. Fixed build issues.

* Fixed gcc issues

* Extracted block producer from table scan operator

* Block producer is now used in system tables

* More warnings on linux builds

+++ csvsqldb/variant.cpp
@@ -389,7 +389,7 @@ namespace csvsqldb
     if (_type == INT) {
       _storage._int += rhs;
     } else if (_type == REAL) {
-      _storage._real += rhs;
+      _storage._real += static_cast<double>(rhs);
     }
     return *this;
   }
@@ -436,7 +436,7 @@ namespace csvsqldb
         if (rhs._storage._int == 0) {
           CSVSQLDB_THROW(VariantException, "cannot devide by null");
         }
-        _storage._real /= rhs._storage._int;
+        _storage._real /= static_cast<double>(rhs._storage._int);
       } else {
         if (Approx(rhs._storage._real) == 0.0) {
           CSVSQLDB_THROW(VariantException, "cannot devide by null");
